fixed references in graphql, needs testing

+++ server/src/modules/category/category.resolver.ts
@@ -1,34 +1,72 @@
-import { Args, Mutation, Query, Resolver } from '@nestjs/graphql';
-import { CategoryService } from './category.service';
-import { Category } from './entities/category.entity';
-import { CreateCategoryInput } from './dto/create-category.input';
-import { UpdateCategoryInput } from './dto/update-category.input';
+import {Args, Mutation, Query, Resolver} from '@nestjs/graphql';
+import {CategoryService} from './category.service';
+import {Category} from './entities/category.entity';
+import {CreateCategoryInput} from './dto/create-category.input';
+import {UpdateCategoryInput} from './dto/update-category.input';
+import {CurrentUser} from "../../decorators/auth.decorator";
+import {GraphQLError} from "graphql";
+import {FileUpload, GraphQLUpload} from "graphql-upload";
+import {UseGuards} from "@nestjs/common";
+import {JwtAuthGuard} from "../auth/auth.guard";
+import {SharedService} from "../shared/shared.service";
 
 @Resolver(() => Category)
+@UseGuards(JwtAuthGuard)
 export class CategoryResolver {
-  constructor(private readonly categoryService: CategoryService) {}
+  constructor(
+    private readonly categoryService: CategoryService,
+    private readonly sharedService: SharedService
+    ) {}
 
   @Mutation(() => Category)
-  createCategory(
+  async createCategory(
+    @CurrentUser() user,
     @Args('createCategoryInput') createCategoryInput: CreateCategoryInput,
+    @Args({ name: 'file', type: () => GraphQLUpload })
+      { createReadStream, filename }: FileUpload,
   ) {
-    return this.categoryService.create(createCategoryInput);
+    if (!user) return new GraphQLError("Unauthorized");
+    createCategoryInput.createdBy = user._id;
+    const category = await this.categoryService.create(createCategoryInput);
+    const path = `./uploads/${category._id}`;
+    await this.sharedService.uploadImage(createReadStream, path);
+    return category;
+  }
+
+  @Mutation(() => String)
+  async uploadCategoryImage(
+    @CurrentUser() user,
+    @Args('id', { type: () => String }) id: string,
+    @Args({ name: 'file', type: () => GraphQLUpload })
+      { createReadStream, filename }: FileUpload,
+  ): Promise<any> {
+    if (!user) return new GraphQLError("Unauthorized");
+    const path = `./uploads/${id}`;
+    await this.sharedService.uploadImage(createReadStream, path);
+    return path;
   }
 
   @Query(() => [Category], { name: 'category' })
-  findAll() {
+  findAll(@CurrentUser() user) {
+    if (!user) return new GraphQLError("Unauthorized");
     return this.categoryService.findAll();
   }
 
   @Query(() => Category, { name: 'category' })
-  findOne(@Args('id', { type: () => String }) id: string) {
+  findOne(
+    @CurrentUser() user,
+    @Args('id', { type: () => String }) id: string
+  ) {
+    if (!user) return new GraphQLError("Unauthorized");
     return this.categoryService.findOne(id);
   }
 
   @Mutation(() => Category)
   updateCategory(
+    @CurrentUser() user,
     @Args('updateCategoryInput') updateCategoryInput: UpdateCategoryInput,
   ) {
+    if (!user) return new GraphQLError("Unauthorized");
     return this.categoryService.update(
       updateCategoryInput._id,
       updateCategoryInput,
@@ -36,7 +74,11 @@ export class CategoryResolver {
   }
 
   @Mutation(() => Category)
-  removeCategory(@Args('id', { type: () => String }) id: string) {
+  removeCategory(
+    @CurrentUser() user,
+    @Args('id', { type: () => String }) id: string
+  ) {
+    if (!user) return new GraphQLError("Unauthorized");
     return this.categoryService.remove(id);
   }
 }
